Move interfaces for schematiclayer into its own
